requirements_agentcore.txt
bedrock-agentcore-runtime>=1.0.0
bedrock-agentcore-memory>=1.0.0
bedrock-agentcore-tools>=1.0.0
boto3>=1.34.0
pydantic>=2.0.0
python-dateutil>=2.8.0
PyMySQL>=1.1.0